fix : AuthenticationPrincipal 사용하여 본인만 리뷰 수정/삭제 가능

See merge request cloud_track/class_03/web_project3/team05/hospital_here_spring!10

+++ hospital_here/hospital_here/src/main/java/com/team5/hospital_here/review/service/ReviewService.java
@@ -1,5 +1,8 @@
 package com.team5.hospital_here.review.service;
 
+import com.team5.hospital_here.common.exception.CustomException;
+import com.team5.hospital_here.common.exception.ErrorCode;
+import com.team5.hospital_here.common.jwt.CustomUser;
 import com.team5.hospital_here.hospital.entity.Hospital;
 import com.team5.hospital_here.hospital.repository.HospitalRepository;
 import com.team5.hospital_here.review.entity.ReviewDTO;
@@ -8,13 +11,16 @@ import com.team5.hospital_here.review.entity.ReviewMapper;
 import com.team5.hospital_here.review.repository.ReviewRepository;
 import com.team5.hospital_here.user.entity.user.User;
 import com.team5.hospital_here.user.repository.UserRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 public class ReviewService {
 
@@ -47,18 +53,48 @@ public class ReviewService {
         reviewRepository.save(reviewEntity);
     }
 
-    public void deleteReview(Long reviewId) {
-        reviewRepository.deleteById(reviewId);
+    public void deleteReview(CustomUser customUser, Long reviewId) {
+
+        Optional<ReviewEntity> review = reviewRepository.findById(reviewId);
+        Long userId = review.get().getUser().getId();
+        log.info("접속중 유저 아이디 : {}",customUser.getUser().getId());
+        log.info("삭제할 리뷰 유저 아이디: {}",userId);
+        if(userId.equals(customUser.getUser().getId()))
+        {
+            reviewRepository.deleteById(reviewId);
+        }
+        else {
+            throw new CustomException(ErrorCode.NO_PERMISSION);
+        }
+
     }
 
-    public void updateReview(Long reviewId, ReviewDTO reviewDTO) {
+    public void updateReview(CustomUser customUser, Long reviewId, ReviewDTO reviewDTO) {
 
-        ReviewEntity reviewEntity = reviewRepository.findById(reviewId).orElse(null);
+        log.info("유저 아이디 :{}",customUser.getUser().getId());
+        log.info("리뷰적은 유저 아이디:{}",reviewRepository.findById(reviewId).get().getUser().getId());
+        if(customUser.getUser().getId().equals(reviewRepository.findById(reviewId).get().getUser().getId())){
+            ReviewEntity reviewEntity = reviewRepository.findById(reviewId).orElse(null);
 
-        reviewEntity.setContent(reviewDTO.getContent());
-        reviewEntity.setRating(reviewDTO.getRating());
+            reviewEntity.setContent(reviewDTO.getContent());
+            reviewEntity.setRating(reviewDTO.getRating());
 
-        reviewRepository.save(reviewEntity);
+            reviewRepository.save(reviewEntity);
+        }
+        else {
+            throw new CustomException(ErrorCode.NO_PERMISSION);
+        }
+
+    }
+
+    public List<ReviewDTO> findByHospitalId(Long hospitalId)
+    {
+        List<ReviewEntity> reviewEntities = reviewRepository.findByHospitalId(hospitalId);
+
+
+        return reviewEntities.stream()
+                .map(ReviewMapper::toReviewDTO)
+                .toList();
     }
